lab12/requirements.txt
fastmcp
requests
streamlit
langchain-core
langchain-google-genai
google-genai
